remove i18n import from a bunch of templates
these were needed originally because of a hack (spline.lib.makoext)
that rewrote the template to inject the line `_ = i18n.Translator()`.
this was necessary because of bothersome details about making
translations work across plugins.

in the pyramid port, we don't have to deal with plugins, so we can just
do it the normal way by setting up _ as a renderer global. so these
imports aren't needed any more.