Enable sessions and make obdurate work
Gonna have to pull beaker in for caching anyway, so might as well use it
for sessions. I'd rather do without sessions, but i don't want to go
through the trouble of figuring out something different right now.

While we're at it, set the minimum version for pyramid_beaker to 0.8 (the
latest release, from 2013).

Updates #115