Implement flash for lookup misspellings
This commit ports the Flash class from spline, with a couple minor
modifications to deal with differences in the pyramid version. The
session now has to be passed explicitly to the constructor, and
pop_messages was changed slightly to deal with autosaving.

This allows lookup to once again berate the user for minor misspellings,
which was the only uses of flash in spline-pokedex.