Get caching working (first pass)
pyramid_beaker does pretty much nothing for us here. It configures cache
regions, but we don't actually use them, so we have to set everything up
ourselves.

This (re-)adds caching to the pokemon page. Other pages that used to be
cached are now broken, but i'll fix them in upcoming commits.

Updates #128