hook up error view
Note that the error view does not trigger for HTTP exceptions which are
*returned* from their views rather than *raised*. I didn't know that
until now, so that's what all the existing views do - return
HTTPNotFound when a resource isn't found instead of raising it.

    e.g. /dex/pokemon/notfound

The not found handler does get called in other circumstances, such as
when a url fails to map to any valid route.

    e.g. /dex/digimon

I'm leaving this as-is for now since we might want to show a different
error page in the former case.