show pretty not found messages for missing resources
Show veekun's error page when a pokemon or other resource isn't found,
instead of pyramid's generic one. Note that errors which are returned
from views rather than raised are not caught by error views.

While we're at it, leave room for showing a more detailed error message
in the future.