frontpage: remove cron stuff
The original frontpage design required a cron job to hit the /cron
endpoint every minute or so, which would periodically run background
tasks like updating the frontpage sources.

Remove that mess and just rely on cache expiration.

The one nice thing about the old design was that updates happened in the
background, not in a user request; it would be bad if visiting the
frontpage could kick off dozens of (potentially slow) feed requests.
But that's solving a problem we never really had. We don't have many
feeds. One is eevee's blog, which is hosted on the same server, and the
other is the git changelog, which just needs local file access. (Also
Bulbanews, but they haven't posted anything in a while.) So it seems
better to avoid that complexity.

Advantages:
- don't need to wait an hour for feed items to show up
- no unnecessary updates if the site is idle
- less complex

Disadvantages:
- moves feed updates into the request path